migrate all eod related files to parent eod module
