+++ src/api/endpoint_prelude.rs
@@ -1,8 +1,3 @@
-// Licensed under the MIT license
-// <LICENSE-MIT or http://opensource.org/licenses/MIT>.
-// This file may not be copied, modified, or distributed
-// except according to those terms.
-
 //! Endpoint prelude
 //!
 //! This module re-exports all of the types needed for endpoints to implement the
