+++ src/auth.rs
@@ -1,7 +1,3 @@
-// Licensed under the MIT license
-// <LICENSE-MIT or http://opensource.org/licenses/MIT>.
-// This file may not be copied, modified, or distributed
-// except according to those terms.
 use thiserror::Error;
 
 use crate::api::{self, Query};
