+++ src/test/client.rs
@@ -1,8 +1,3 @@
-// Licensed under the MIT license
-// <LICENSE-MIT or http://opensource.org/licenses/MIT>.
-// This file may not be copied, modified, or distributed
-// except according to those terms.
-
 use std::borrow::Cow;
 use std::cmp;
 use std::collections::HashMap;
